Canvas: Move all items from Syllabus sub panel and place them on Course Home panel [#1274]
