+++ lib/ui/canvas/CanvasCourseHomePanel.dart
@@ -18,6 +18,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/model/Canvas.dart';
 import 'package:illinois/ui/canvas/CanvasAccountNotificationsPanel.dart';
+import 'package:illinois/ui/canvas/CanvasCourseModulesPanel.dart';
+import 'package:illinois/ui/canvas/CanvasSyllabusHtmlPanel.dart';
 import 'package:rokwire_plugin/model/group.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Canvas.dart';
@@ -25,7 +27,6 @@ import 'package:rokwire_plugin/service/groups.dart';
 import 'package:illinois/ui/canvas/CanvasCourseAnnouncementsPanel.dart';
 import 'package:illinois/ui/canvas/CanvasCourseCalendarPanel.dart';
 import 'package:illinois/ui/canvas/CanvasCourseCollaborationsPanel.dart';
-import 'package:illinois/ui/canvas/CanvasCourseSyllabusPanel.dart';
 import 'package:illinois/ui/canvas/CanvasFileSystemEntitiesListPanel.dart';
 import 'package:illinois/ui/canvas/CanvasWidgets.dart';
 import 'package:illinois/ui/groups/GroupWidgets.dart';
@@ -116,7 +117,7 @@ class _CanvasCourseHomePanelState extends State<CanvasCourseHomePanel> {
       RibbonButton(
           label: Localization().getStringEx('panel.home_canvas_course.button.syllabus.title', 'Syllabus'),
           hint: Localization().getStringEx('panel.home_canvas_course.button.syllabus.hint', ''),
-          leftIcon: 'images/icon-canvas-partially-implemented.png',
+          leftIcon: 'images/icon-canvas-implemented-working.png',
           onTap: _onTapSyllabus),
       _buildDelimiter(),
       RibbonButton(
@@ -153,8 +154,28 @@ class _CanvasCourseHomePanelState extends State<CanvasCourseHomePanel> {
           label: Localization().getStringEx('panel.home_canvas_course.button.notitification.history.title', 'Notification history'),
           hint: Localization().getStringEx('panel.home_canvas_course.button.notitification.history.hint', ''),
           leftIcon: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapInbox),
+          onTap: _onTapNotificationHistory),
       _buildDelimiter(),
+      RibbonButton(
+          label: Localization().getStringEx('panel.home_canvas_course.button.modules.title', 'Modules'),
+          hint: Localization().getStringEx('panel.home_canvas_course.button.modules.hint', ''),
+          leftIcon: 'images/icon-canvas-implemented-working.png',
+          onTap: _onTapModules),
+      RibbonButton(
+          label: Localization().getStringEx('panel.home_canvas_course.button.assignments.title', 'Assignments'),
+          hint: Localization().getStringEx('panel.home_canvas_course.button.assignments.hint', ''),
+          leftIcon: 'images/icon-settings.png'),
+      _buildDelimiter(),
+      RibbonButton(
+          label: Localization().getStringEx('panel.home_canvas_course.button.zoom.title', 'Zoom Meeting'),
+          hint: Localization().getStringEx('panel.home_canvas_course.button.zoom.hint', ''),
+          leftIcon: 'images/icon-settings.png'),
+      _buildDelimiter(),
+      RibbonButton(
+          label: Localization().getStringEx('panel.home_canvas_course.button.grades.title', 'Grades'),
+          hint: Localization().getStringEx('panel.home_canvas_course.button.grades.hint', ''),
+          leftIcon: 'images/icon-settings.png'),
+      _buildDelimiter()
     ]);
   }
 
@@ -163,38 +184,43 @@ class _CanvasCourseHomePanelState extends State<CanvasCourseHomePanel> {
   }
 
   void _onTapSyllabus() {
-    Analytics().logSelect(target: "Canvas Course -> Syllabus");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseSyllabusPanel(courseId: widget.courseId!)));
+    Analytics().logSelect(target: 'Canvas Course -> Syllabus');
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasSyllabusHtmlPanel(courseId: widget.courseId!)));
   }
 
   void _onTapAnnouncements() {
-    Analytics().logSelect(target: "Canvas Course -> Announcements");
+    Analytics().logSelect(target: 'Canvas Course -> Announcements');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseAnnouncementsPanel(courseId: widget.courseId!)));
   }
 
   void _onTapFiles() {
-    Analytics().logSelect(target: "Canvas Course -> Files");
+    Analytics().logSelect(target: 'Canvas Course -> Files');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasFileSystemEntitiesListPanel(courseId: widget.courseId)));
   }
 
   void _onTapCollaborations() {
-    Analytics().logSelect(target: "Canvas Course -> Collaborations");
+    Analytics().logSelect(target: 'Canvas Course -> Collaborations');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseCollaborationsPanel(courseId: widget.courseId!)));
   }
 
   void _onTapCalendar() {
-    Analytics().logSelect(target: "Canvas Course -> Calendar");
+    Analytics().logSelect(target: 'Canvas Course -> Calendar');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseCalendarPanel(courseId: widget.courseId!)));
   }
 
   void _onTapNotifications() {
-    Analytics().logSelect(target: "Canvas Course -> Notifications");
+    Analytics().logSelect(target: 'Canvas Course -> Notifications');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasAccountNotificationsPanel()));
   }
 
-  void _onTapInbox() {
-    Analytics().logSelect(target: "Canvas Course -> Inbox");
+  void _onTapNotificationHistory() {
+    Analytics().logSelect(target: 'Canvas Course -> Notification history');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => InboxHomePanel()));
+  }
+
+  void _onTapModules() {
+    Analytics().logSelect(target: 'Canvas Course -> Modules');
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseModulesPanel(courseId: widget.courseId!)));
   }
 
   void _loadCourse() {
